requirements.txt
sentence-transformers==3.4.1
chromadb==0.6.3
anthropic==0.50.0
pypdf==5.4.0


annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.13.0
asgiref==3.11.1
backoff==2.2.1
bcrypt==5.0.0
build==1.4.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.3.2
cryptography==46.0.7
distro==1.9.0
durationpy==0.10
fastapi==0.136.0
filelock==3.29.0
flatbuffers==25.12.19
fsspec==2026.3.0
google-auth==2.49.2
google-genai==1.73.1
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==7.1.0
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
kubernetes==35.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
oauthlib==3.3.1
onnxruntime==1.24.4
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-instrumentation==0.62b0
opentelemetry-instrumentation-asgi==0.62b0
opentelemetry-instrumentation-fastapi==0.62b0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opentelemetry-util-http==0.62b0
orjson==3.11.8
overrides==7.7.0
packaging==26.1
pillow==12.2.0
posthog==7.12.0
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pypdf==5.4.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==3.4.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==4.57.6
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.2
zipp==3.23.1
